Added React-Router for routing
Separated components into different jsx files. Added 1 dummy page
component "AboutPage". Use ReactRouter for browser routing.
