+++ app/components/news_list.jsx
@@ -1,0 +1,28 @@
+
+'use strict'
+
+var React = require('react');
+var NewsListItem = require('./news_list_item');
+
+// extract id from a news item
+var newsId = function (item) {
+  return item._id.$oid;
+}
+
+// NewsList
+module.exports = React.createClass({
+  render: function() {
+    var newsItems = this.props.items.map(function(data, index) {
+      return (
+        <NewsListItem
+          key={newsId(data)}
+          data={data} />
+      );
+    });
+    return (
+      <div className="news-list">
+        {newsItems}
+      </div>      
+    );
+  }
+});
