+++ package.json
@@ -18,6 +18,7 @@
 		"prompt": "^0.2.14",
 		"jsx-loader": "^0.13.2",
 		"react": "^0.13.3",
+		"react-router": "^0.13.3",
 		"webpack": "^1.10.1"
 	}
 }
